Select CSV data and find line indices
This script highlights some useful functions in the "pandas" library for Python 3 that may help researchers locate specific data sets in their CSV files. The "index" function is used to retrieve the index or ID of CSV table rows in which the selected data are stored. Please note that the final result can be obtained much faster. The purpose of this script is, however, to demonstrate individual operations that may need to be performed independently in other circumstances.