Python for digital history -- recommendations
For further information, you can read the Python documentation for individual packages on [PyPi][https://pypi.org/]. Please keep in mind that Python, like any other software, is constantly being updated and that all packages listed are meant to be used with the latest release, Python3. It is also advisable to check out DH-related repositories on GitHub (https://github.com/), the leading open source code-sharing platform, discussions in coding-forums such as Stack Overflow (https://stackoverflow.com/), or follow professionally curated academics blogs such “The Programming Historian” (https://programminghistorian.org/). As you will see in the table, several popular packages are already included in the latest release of Python and or come with the Anaconda platform (https://www.anaconda.com/distribution/). 
If you have not installed Python yet, we recommend installing Anaconda, which includes Python, to make sure that you are not creating potentially conflicting environments. Anaconda is an easy-to-use environment, especially for people with little knowledge of programming.In order to invoke a particular Python package or a module from that package in your script, you need to type “import” or “import … from” at the beginning. The “import” statement in Python is explained in the official documentation (https://docs.python.org/3/reference/import.html).
Whatever your research goals are, do not hesitate to contact the global community of Python users and developers in one of the many Python forums or via Twitter. For coding women, for examle, there is the fantastic global network [@pyladies] [https://twitter.com/pyladies]. You will always find help.